feat: rename interface to settings
