+++ src/views/Settings.vue
